fix(ci): first-party file line limit — split oversized part_011 test module
